+++ packages/dicomImageLoader/src/imageLoader/index.ts
@@ -17,6 +17,11 @@ import { default as getMinMax } from '../shared/getMinMax';
 import { default as isColorImage } from '../shared/isColorImage';
 import { default as isJPEGBaseline8BitColor } from './isJPEGBaseline8BitColor';
 import { default as getPixelData } from './wadors/getPixelData';
+import { default as getScalingParameters } from './getScalingParameters';
+import { default as isColorConversionRequired } from './isColorConversionRequired';
+import { default as removeAFromRGBA } from './removeAFromRGBA';
+import { default as isModalityLUTForDisplay } from './isModalityLutForDisplay';
+import { default as setPixelDataType } from './setPixelDataType';
 import { internal } from './internal/index';
 
 const cornerstoneDICOMImageLoader = {
@@ -36,6 +41,11 @@ const cornerstoneDICOMImageLoader = {
   getMinMax,
   isColorImage,
   isJPEGBaseline8BitColor,
+  getScalingParameters,
+  isColorConversionRequired,
+  removeAFromRGBA,
+  isModalityLUTForDisplay,
+  setPixelDataType,
   internal,
 };
