+++ packages/dicomImageLoader/src/imageLoader/isModalityLutForDisplay.ts
@@ -1,0 +1,10 @@
+function isModalityLUTForDisplay(sopClassUid: string): boolean {
+  // special case for XA and XRF
+  // https://groups.google.com/forum/#!searchin/comp.protocols.dicom/Modality$20LUT$20XA/comp.protocols.dicom/UBxhOZ2anJ0/D0R_QP8V2wIJ
+  return (
+    sopClassUid !== '1.2.840.10008.5.1.4.1.1.12.1' && // XA
+    sopClassUid !== '1.2.840.10008.5.1.4.1.1.12.2.1'
+  ); // XRF
+}
+
+export default isModalityLUTForDisplay;
